Retry delete pod deployment in test.
Summary:
`MesosAppIntegrationTest.MesosApp should deploy a simple pod with health checks`
sometimes fails with "409 was not equal to 202" which means the delete
has a conflict. To avoid this error we just retry the delete.

I did not verify why there is another deployment running. I assume that
there is a slight race condition: We receive the deployment finished
event but the deployment is not completely removed yet.

Test Plan: pipeline

Reviewers: zen-dog, timcharper

Subscribers: marathon-dev, marathon-team, jenkins

JIRA Issues: MARATHON-7467

Differential Revision: https://phabricator.mesosphere.com/D833