Backup/Restore is not working correctly if an app was deleted before
Summary:
In the implementation to do the backup `future.filter(..)` was used, which produces an exception if future given filter condition is false.
If no app definition is returned for the queried appId, false is returned and therefore startup of marathon is aborted. This fix deletes the usage
of `future.filter(..)`

Test Plan: sbt "integration:test-only *DeleteAppAndBackupIntegrationTest"

Reviewers: zen-dog, jeschkies, timcharper

Subscribers: marathon-dev, marathon-team

JIRA Issues: MARATHON-7525

Differential Revision: https://phabricator.mesosphere.com/D862